start using 8/16/32 bits registers 1. compiler knows all the registers 2. but only 64bits register is a color (for reg alloc) 3. backend records the length of GPR for each operand during instruction selection 4. after reg alloc, when replacing temp with a color, find corresponding GPR for the length recorded before

+++ tests/test_compiler/test_controlflow.rs
@@ -330,7 +330,7 @@ fn test_sgt_value() {
     let lib = testutil::compile_fnc("sgt_value", &sgt_value);
 
     unsafe {
-        let sgt_value : libloading::Symbol<unsafe extern fn(i64, i64) -> u64> = lib.get(b"sgt_value").unwrap();
+        let sgt_value : libloading::Symbol<unsafe extern fn(i64, i64) -> u8> = lib.get(b"sgt_value").unwrap();
 
         let res = sgt_value(255, 0);
         println!("sgt_value(255, 0) = {}", res);
@@ -386,7 +386,7 @@ fn test_sgt_u8_value() {
     let lib = testutil::compile_fnc("sgt_u8_value", &sgt_u8_value);
 
     unsafe {
-        let sgt_u8_value : libloading::Symbol<unsafe extern fn(i8, i8) -> u64> = lib.get(b"sgt_u8_value").unwrap();
+        let sgt_u8_value : libloading::Symbol<unsafe extern fn(i8, i8) -> u8> = lib.get(b"sgt_u8_value").unwrap();
 
         let res = sgt_u8_value(-1, 0);
         println!("sgt_u8_value(-1, 0) = {}", res);
